Gravity brings everyone together for the holidays.
Apologies for the Octave code, but I have too much to do today to mess with Python. The port should be quick for the fluent....

The parameters chosen here appear to yield relative stability, though I doubt that the code conserves energy. If execution speed is a problem, RK4 should allow a dramatic increase in timestep. Loop rate in Octave on a 3 GHz i5-3330 is 5kHz, so I'm hopeful that the RPi can hang with it at 24 Hz.